tests: Move telemetry_messenger_spy to infection_monkey/conftest.py
