+++ monkey/tests/unit_tests/infection_monkey/conftest.py
@@ -1,0 +1,17 @@
+import pytest
+
+from infection_monkey.telemetry.i_telem import ITelem
+from infection_monkey.telemetry.messengers.i_telemetry_messenger import ITelemetryMessenger
+
+
+class TelemetryMessengerSpy(ITelemetryMessenger):
+    def __init__(self):
+        self.telemetries = []
+
+    def send_telemetry(self, telemetry: ITelem):
+        self.telemetries.append(telemetry)
+
+
+@pytest.fixture
+def telemetry_messenger_spy():
+    return TelemetryMessengerSpy()
